+++ server/src/main/kotlin/com/adobe/testing/s3mock/store/BucketMetadata.kt
@@ -21,6 +21,7 @@ import com.adobe.testing.s3mock.dto.BucketLifecycleConfiguration
 import com.adobe.testing.s3mock.dto.LocationInfo
 import com.adobe.testing.s3mock.dto.ObjectLockConfiguration
 import com.adobe.testing.s3mock.dto.ObjectOwnership
+import com.adobe.testing.s3mock.dto.ServerSideEncryptionConfiguration
 import com.adobe.testing.s3mock.dto.VersioningConfiguration
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.fasterxml.jackson.annotation.JsonProperty
@@ -41,6 +42,7 @@ data class BucketMetadata(
   val bucketRegion: String,
   val bucketInfo: BucketInfo?,
   val locationInfo: LocationInfo?,
+  val bucketEncryptionConfiguration: ServerSideEncryptionConfiguration? = null,
   @param:JsonProperty("objects")
   private val _objects: MutableMap<String, UUID> = mutableMapOf(),
 ) {
@@ -67,6 +69,9 @@ data class BucketMetadata(
 
   fun withBucketLifecycleConfiguration(bucketLifecycleConfiguration: BucketLifecycleConfiguration?): BucketMetadata =
     this.copy(bucketLifecycleConfiguration = bucketLifecycleConfiguration)
+
+  fun withBucketEncryptionConfiguration(bucketEncryptionConfiguration: ServerSideEncryptionConfiguration?): BucketMetadata =
+    this.copy(bucketEncryptionConfiguration = bucketEncryptionConfiguration)
 
   @get:JsonIgnore
   val isVersioningEnabled: Boolean